separate pagination logic and error handling into distinct files
also add pagination method GetAllPages
